+++ packages/ice/src/service/config.ts
@@ -1,7 +1,6 @@
 import * as path from 'path';
 import fs from 'fs-extra';
 import type { ServerCompiler } from '../types/plugin.js';
-import removeTopLevelCode from '../esbuild/removeTopLevelCode.js';
 import { getCache, setCache } from '../utils/persistentCache.js';
 import { getFileHash } from '../utils/hash.js';
 import dynamicImport from '../utils/dynamicImport.js';
@@ -15,7 +14,7 @@ type GetOutfile = (entry: string, exportNames: string[]) => string;
 interface CompileConfig {
   entry: string;
   rootDir: string;
-  transformInclude: (id: string) => boolean;
+  transformInclude?: (id: string) => boolean;
   needRecompile?: (entry: string, options: string[]) => Promise<boolean | string>;
   getOutfile?: GetOutfile;
 }
@@ -48,13 +47,19 @@ class Config {
         format: 'esm',
         outfile,
         plugins: [
-          removeTopLevelCode(keepExports, transformInclude),
           // External node builtin modules, such as `fs`, it will be imported by weex document.
           externalBuiltinPlugin(),
-        ],
+        ].filter(Boolean),
         sourcemap: false,
         logLevel: 'silent', // The main server compiler process will log it.
-      }, {});
+      }, {
+        swc: {
+          keepExports: {
+            value: keepExports,
+            include: transformInclude,
+          },
+        },
+      });
       if (!error) {
         this.status = 'RESOLVED';
         return outfile;
@@ -197,8 +202,6 @@ export const getRouteExportConfig = (rootDir: string) => {
     entry: routeConfigFile,
     rootDir,
     getOutfile: getRouteConfigOutfile,
-    // Only remove top level code for route component file.
-    transformInclude: (id) => id.includes('src/pages'),
     needRecompile: async (entry) => {
       let cached = false;
       try {
@@ -218,8 +221,6 @@ export const getRouteExportConfig = (rootDir: string) => {
     entry: loadersConfigFile,
     rootDir,
     getOutfile: getdataLoadersConfigOutfile,
-    // Only remove top level code for route component file.
-    transformInclude: (id) => id.includes('src/pages'),
     needRecompile: async (entry) => {
       let cached = false;
       const cachedKey = `loader_config_file_${process.env.__ICE_VERSION__}`;