+++ packages/ice/tests/fixtures/removeCode/import.ts
@@ -1,8 +1,0 @@
-import { a, b } from 'test';
-import { a as c } from 'test-a';
-import d from 'test-d';
-import 'test-c';
-
-export function pageConfig() {
-  return { a: 1 };
-}