+++ packages/miniapp-runtime/src/app/index.ts
@@ -1,8 +1,10 @@
-import { getAppConfig, defineAppConfig, useAppData, useData, useConfig, defineDataLoader } from '@ice/runtime';
+import { getAppConfig, defineAppConfig, useAppData, defineDataLoader } from '@ice/runtime';
 import runClientApp from './runClientApp.js';
 import Link from './Link.js';
 import useSearchParams from './useSearchParams.js';
 import { routerHistory as history } from './history.js';
+import usePageLifecycle from './usePageLifecycle.js';
+import { useData, useConfig } from './routeContext.js';
 
 export {
   runClientApp,
@@ -15,4 +17,5 @@ export {
   useSearchParams,
   history,
   defineDataLoader,
+  usePageLifecycle,
 };