+++ packages/plugin-i18n/src/typings.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@ice/pkg/types" />