+++ packages/rax-compat/src/possible-standard-names.ts
@@ -1,20 +1,35 @@
 // Select part from https://github.com/facebook/react/blob/main/packages/react-dom/src/shared/possibleStandardNames.js#L11
-const possibleStandardNames: Record<string, string> = {
-  autofocus: 'autoFocus',
-  autoplay: 'autoPlay',
-  classname: 'className',
-  crossorigin: 'crossOrigin',
-  defaultchecked: 'defaultChecked',
-  defaultvalue: 'defaultValue',
+const possibleStandardNames = [
+  // Common
+  'className',
+  'innerHTML',
+  'autoFocus',
+  'inputMode',
+  // script
+  'crossOrigin',
+  // input
+  'autoPlay',
+  'autoComplete',
+  'defaultChecked',
+  'defaultValue',
+  'htmlFor',
+  'maxLength',
+  'minLength',
+  'readOnly',
+  // video
+  'playsInline',
+  // frame
+  'marginWidth',
+  'marginHeight',
+  // meta
+  'charSet',
+].reduce((records: Record<string, string>, iter: string) => {
+  records[iter.toLowerCase()] = iter;
+  return records;
+}, {
+  // Special cases.
+  class: 'className',
   for: 'htmlFor',
-  htmlfor: 'htmlFor',
-  innerhtml: 'innerHTML',
-  inputmode: 'inputMode',
-  marginwidth: 'marginWidth',
-  marginheight: 'marginHeight',
-  maxlength: 'maxLength',
-  minlength: 'minLength',
-  readonly: 'readOnly',
-};
+});
 
 export default possibleStandardNames;