+++ packages/runtime/src/usePageLifecycle.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+export default function usePageLifecycle(lifecycle: string, callback: any) {
+  // Just compatible with miniapp
+  return;
+}